+++ src/bmi.jl
@@ -76,7 +76,10 @@ end
 "Write state output to netCDF and close files."
 function BMI.finalize(model::Model)
     @unpack config, writer, clock = model
-    write_netcdf_timestep(model, writer.state_dataset, writer.state_parameters)
+    # it is possible that the state dataset has been closed by `save_state`  
+    if !isnothing(writer.state_dataset) && isopen(writer.state_dataset)
+        write_netcdf_timestep(model, writer.state_dataset, writer.state_parameters)
+    end
     reset_clock!(model.clock, config)
     close_files(model, delete_output = false)
 end
@@ -394,6 +397,7 @@ function save_state(model::Model)
         @info "Write output states to netCDF file `$(model.writer.state_nc_path)`."
     end
     write_netcdf_timestep(model, writer.state_dataset, writer.state_parameters)
+    close(writer.state_dataset)
 end
 
 function get_start_unix_time(model::Model)